forge/templates/gatekeeper/pyproject.toml
[project]
name = "gatekeeper"
dynamic = ["version"]
description = "OIDC Relying Party and Traefik ForwardAuth gateway for multi-tenant authentication"
authors = [
    {name = "Strive", email = "dev@strive.us"},
]
dependencies = [
    "typer>=0.15.1",
    "uvicorn>=0.34.0",
    "colorlog>=6.9.0",
    "fastapi>=0.115.12",
    "pyyaml>=6.0.3",
    "python-dotenv>=1.1.0",
    "pydantic>=2.11.1",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.20",
    "PyJWT[crypto]>=2.10.1",
    "httpx>=0.27.0",
    "redis>=5.0.0",
    "prometheus-client>=0.21.0",
]
requires-python = ">=3.13"
readme = "README.md"
license = {text = "Proprietary"}
version = "0.1.0"


[tool.pdm]
distribution = false

[tool.pdm.build]
includes = ["src"]

[tool.pdm.scripts]
lint = "pre-commit run --all-files"
check = "ruff check ."
fmt = "ruff format ."
docs = "mkdocs serve"
typecheck = "ty check src/"
tests = "pytest -v tests"
coverage = "pytest --cov --cov-report=term-missing --cov-report=html tests"

[tool.pdm.version]
source = "scm"

[tool.ruff.lint]
select = [
    "B", # flake8-bugbear
    "C4", # flake8-comprehensions
    "E", # pycodestyle - Error
    "F", # Pyflakes
    "I", # isort
    "W", # pycodestyle - Warning
    "UP", # pyupgrade
]
ignore = [
    "E501", # line-too-long
    "W191", # tab-indentation
    "I001", # isort - import order
]

[tool.ruff.lint.isort]
known-first-party = ["gatekeeper"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = ["fastapi.Depends", "fastapi.params.Depends", "fastapi.Query", "fastapi.params.Query"]

[tool.ty.environment]
python-version = "3.13"
extra-paths = ["src"]

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = ["tests"]
addopts = "-ra -q"
asyncio_mode = "auto"
markers = [
    "container: tests requiring Docker (Testcontainers)",
]

[tool.coverage.run]
parallel = true
source = ["src"]
omit = [
    "src/app/__main__.py",
    "src/app/cli/*",
]

[tool.coverage.report]
fail_under = 60
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.",
    "if TYPE_CHECKING:",
]

[tool.coverage.html]
directory = "htmlcov"

[dependency-groups]
dev = [
    "ty>=0.0.1a11",
]
lint = [
    "ruff>=0.3.4",
    "pre-commit>=3.5.0",
]
docs = [
    "mkdocs>=1.5.3",
    "mkdocs-material>=9.5.15",
    "mkdocstrings[python]>=0.24.1",
]
test = [
    "pytest>=8.1.1",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "fakeredis>=2.21.0",
    "testcontainers[redis]>=4.0.0",
]
